church_backend/requirements.txt
flask
sqlalchemy
flask-migrate
python-dotenv
psycopg2-binary
flask-jwt-extended
flask-bcrypt 
flask-dance 